+++ tests/synopse/conftest.py
@@ -23,10 +23,10 @@ from mex.extractors.primary_source.helpers import (
     get_extracted_primary_source_id_by_name,
 )
 from mex.extractors.settings import Settings
-from mex.extractors.synopse.models.project import SynopseProject
-from mex.extractors.synopse.models.study import SynopseStudy
-from mex.extractors.synopse.models.study_overview import SynopseStudyOverview
-from mex.extractors.synopse.models.variable import SynopseVariable
+from mex.extractors.synopse.models.project import ProjektUndStudienverwaltung
+from mex.extractors.synopse.models.study import MetadatenZuDatensaetzen
+from mex.extractors.synopse.models.study_overview import Datensatzuebersicht
+from mex.extractors.synopse.models.variable import Variablenuebersicht
 from mex.extractors.synopse.transform import (
     transform_synopse_variables_to_mex_variable_groups,
 )
@@ -39,74 +39,71 @@ def synopse_variables_raw() -> list[dict[str, str | int | float | None]]:
     return [
         {  # var 1, auspraegung 1
             "textbox49": -97,
-            "Originalfrage": None,
-            "StudieID1": "STUDY1",
-            "StudieID2": 12345,
-            "SymopseID": 1,
+            "originalfrage": None,
+            "studie_id2": 12345,
+            "symopse_id": 1,
             "textbox51": "Nicht erhoben",
             "textbox5": "Krankheiten (1101)",
             "textbox2": "Krankheiten allgemein (110100)",
-            "valInstrument": "Health Questionnaire , Frage 18",
+            "val_instrument": "Health Questionnaire , Frage 18",
             "textbox21": "Angeborene Fehlbildung",
             "textbox24": "KHEfehlb",
             "textbox11": "Zahl",
-            "IntVar": False,
+            "int_var": False,
             "KeepVarname": False,
         },
         {  # var 1, auspraegung 2
             "textbox49": -98,
-            "Originalfrage": None,
-            "StudieID1": "STUDY1",
-            "StudieID2": 12345,
-            "SymopseID": 1,
+            "originalfrage": None,
+            "studie_id2": 12345,
+            "symopse_id": 1,
             "textbox51": "Weiß nicht",
             "textbox5": "Krankheiten (1101)",
             "textbox2": "Krankheiten allgemein (110100)",
-            "valInstrument": None,
+            "val_instrument": None,
             "textbox21": "Angeborene Fehlbildung",
             "textbox24": "KHEfehlb",
             "textbox11": "Text",
-            "IntVar": True,
+            "int_var": True,
             "KeepVarname": False,
         },
-        {  # var 2, missing var label, valInstrument
+        {  # var 2, missing var label, val_instrument
             "textbox49": 1,
-            "Originalfrage": None,
-            "StudieID1": "STUDY1",
-            "StudieID2": 12345,
-            "SymopseID": 2,
+            "originalfrage": None,
+            "studie_id2": 12345,
+            "symopse_id": 2,
             "textbox51": "Ja",
             "textbox5": "Krankheiten (1101)",
             "textbox2": "Krankheiten allgemein (110100)",
-            "valInstrument": None,
+            "val_instrument": None,
             "textbox21": None,
             "textbox24": "KHEfiebB",
             "textbox11": "Zahl",
-            "IntVar": True,
+            "int_var": True,
             "KeepVarname": False,
         },
     ]
 
 
 @pytest.fixture
-def synopse_study_overviews() -> list[SynopseStudyOverview]:
+def synopse_study_overviews() -> list[Datensatzuebersicht]:
     """Return a list Synopse Study Overviews."""
     return [
-        SynopseStudyOverview(
+        Datensatzuebersicht(
             studien_id="12345",
-            ds_typ_id=17,
+            dstyp_id=17,
             titel_datenset="set1",
             synopse_id="synopse1",
         ),
-        SynopseStudyOverview(
+        Datensatzuebersicht(
             studien_id="studie1",
-            ds_typ_id=18,
+            dstyp_id=18,
             titel_datenset="set2",
             synopse_id="synopse1",
         ),
-        SynopseStudyOverview(
+        Datensatzuebersicht(
             studien_id="studie2",
-            ds_typ_id=32,
+            dstyp_id=32,
             titel_datenset="set2",
             synopse_id="synopse2",
         ),
@@ -150,120 +147,100 @@ def synopse_resources() -> list[ExtractedResource]:
 @pytest.fixture
 def synopse_variables(
     synopse_variables_raw: list[dict[str, str | int]],
-) -> list[SynopseVariable]:
+) -> list[Variablenuebersicht]:
     """Return a list Synopse Variables."""
-    return [SynopseVariable.model_validate(v) for v in synopse_variables_raw]
+    return [Variablenuebersicht.model_validate(v) for v in synopse_variables_raw]
 
 
 @pytest.fixture
 def synopse_variables_by_study_id(
-    synopse_variables: list[SynopseVariable],
-) -> dict[int, list[SynopseVariable]]:
+    synopse_variables: list[Variablenuebersicht],
+) -> dict[int, list[Variablenuebersicht]]:
     """Return a mapping from synopse studie id to the variables with this studie id."""
-    synopse_variables = sorted(synopse_variables, key=lambda v: v.studie_id)
+    synopse_variables = sorted(synopse_variables, key=lambda v: v.studie_id2)
     return {
-        studie_id: list(variables)
-        for studie_id, variables in groupby(
-            synopse_variables, key=lambda v: v.studie_id
+        studie_id2: list(variables)
+        for studie_id2, variables in groupby(
+            synopse_variables, key=lambda v: v.studie_id2
         )
     }
 
 
 @pytest.fixture
 def synopse_variables_by_thema(
-    synopse_variables: list[SynopseVariable],
-) -> dict[str, list[SynopseVariable]]:
+    synopse_variables: list[Variablenuebersicht],
+) -> dict[str, list[Variablenuebersicht]]:
     """Return a mapping from synopse thema to the variables with this thema."""
-    synopse_variables = sorted(
-        synopse_variables, key=lambda v: v.thema_und_fragebogenausschnitt
-    )
+    synopse_variables = sorted(synopse_variables, key=lambda v: v.textbox5)
     return {
         thema: list(variables)
-        for thema, variables in groupby(
-            synopse_variables, key=lambda v: v.thema_und_fragebogenausschnitt
-        )
+        for thema, variables in groupby(synopse_variables, key=lambda v: v.textbox5)
     }
 
 
 @pytest.fixture
-def synopse_studies() -> list[SynopseStudy]:
+def synopse_studies() -> list[MetadatenZuDatensaetzen]:
     """Return a list of Synopse Studies."""
     return [
-        SynopseStudy(
-            beitragende="Jane Doe",
-            beschreibung="ein heikles Unterfangen.",
-            dokumentation="Z:\\foo\\bar",
-            ds_typ_id=17,
-            erstellungs_datum="2022",
-            feld_ende="1971-02",
-            feld_start="1970-01",
-            plattform_adresse="S:\\data",
-            rechte="Niemand darf irgendwas.",
-            schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+        MetadatenZuDatensaetzen(
             studien_id="12345",
+            dstyp_id=17,
             titel_datenset="Titel",
-            Studie="Studie123",
-            zugangsbeschraenkung="protected",
-        ),
-        SynopseStudy(
-            beschreibung="ein zweites heikles Unterfangen.",
-            dokumentation="X:\\foo\\bar",
-            ds_typ_id=16,
-            erstellungs_datum="2017",
-            plattform_adresse="blabli blubb",
-            rechte="Niemand darf irgendwas.",
+            beschreibung="ein heikles Unterfangen.",
             schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+            rechte="Niemand darf irgendwas.",
+            zugangsbeschraenkung="protected",
+            datum_der_letzten_aenderung="2022",
+        ),
+        MetadatenZuDatensaetzen(
             studien_id="123456",
+            dstyp_id=16,
             titel_datenset="Titel 2",
+            beschreibung="ein zweites heikles Unterfangen.",
+            schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+            rechte="Niemand darf irgendwas.",
             zugangsbeschraenkung="open",
+            datum_der_letzten_aenderung="2017",
         ),
-        SynopseStudy(
-            beschreibung="eine study ohne Variablen, Projekt, oder exctractedActivity.",
-            dokumentation="interne Datennutzung",
-            ds_typ_id=16,
-            erstellungs_datum="2017",
-            plattform_adresse="blabli blubb",
-            rechte="Niemand darf irgendwas.",
-            schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+        MetadatenZuDatensaetzen(
             studien_id="123457",
+            dstyp_id=16,
             titel_datenset="Study ohne Referenzen",
-            zugangsbeschraenkung="sensitive",
-        ),
-        SynopseStudy(
             beschreibung="eine study ohne Variablen, Projekt, oder exctractedActivity.",
-            dokumentation="https://asd.def",
-            ds_typ_id=16,
-            erstellungs_datum="2017",
-            plattform_adresse="blabli blubb",
-            rechte="Niemand darf irgendwas.",
             schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+            rechte="Niemand darf irgendwas.",
+            zugangsbeschraenkung="sensitive",
+            datum_der_letzten_aenderung="2017",
+        ),
+        MetadatenZuDatensaetzen(
             studien_id="123458",
+            dstyp_id=16,
             titel_datenset="Study 2 ohne Referenzen",
+            beschreibung="eine study ohne Variablen, Projekt, oder exctractedActivity.",
+            schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+            rechte="Niemand darf irgendwas.",
             zugangsbeschraenkung="protected",
+            datum_der_letzten_aenderung="2017",
         ),
-        SynopseStudy(
-            beschreibung="eine study ohne Variablen, Projekt, oder exctractedActivity.",
-            dokumentation="interne Datennutzung",
-            ds_typ_id=16,
-            erstellungs_datum="2017",
-            plattform_adresse="interne Datennutzung",
-            rechte="Niemand darf irgendwas.",
-            schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+        MetadatenZuDatensaetzen(
             studien_id="123457",
+            dstyp_id=16,
             titel_datenset="Study ohne Referenzen",
-            zugangsbeschraenkung="sensitive",
-        ),
-        SynopseStudy(
             beschreibung="eine study ohne Variablen, Projekt, oder exctractedActivity.",
-            dokumentation="https://asd.def",
-            ds_typ_id=16,
-            erstellungs_datum="2017",
-            plattform_adresse="https://asd.def",
-            rechte="Niemand darf irgendwas.",
             schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+            rechte="Niemand darf irgendwas.",
+            zugangsbeschraenkung="sensitive",
+            datum_der_letzten_aenderung="2018",
+        ),
+        MetadatenZuDatensaetzen(
             studien_id="123458",
+            dstyp_id=16,
             titel_datenset="Study 2 ohne Referenzen",
+            beschreibung="eine study ohne Variablen, Projekt, oder exctractedActivity.",
+            schlagworte_themen="Alkohol, Alter und Geschlecht, Drogen",
+            rechte="Niemand darf irgendwas.",
             zugangsbeschraenkung="open",
+            datum_der_letzten_aenderung="2017",
         ),
     ]
 
@@ -278,34 +255,16 @@ def synopse_access_platform() -> AccessPlatformMapping:
 
 
 @pytest.fixture
-def created_by_study_id(synopse_studies: list[SynopseStudy]) -> dict[str, str]:
-    """Return a lookup from study ID to created string."""
-    return {
-        s.studien_id: s.erstellungs_datum
-        for s in synopse_studies
-        if s.erstellungs_datum
-    }
-
-
-@pytest.fixture
-def description_by_study_id(synopse_studies: list[SynopseStudy]) -> dict[str, str]:
+def description_by_study_id(
+    synopse_studies: list[MetadatenZuDatensaetzen],
+) -> dict[str, str]:
     """Return a lookup from study ID to description string."""
     return {s.studien_id: s.beschreibung for s in synopse_studies if s.beschreibung}
 
 
 @pytest.fixture
-def documentation_by_study_id(synopse_studies: list[SynopseStudy]) -> dict[str, Link]:
-    """Return a lookup from study ID to documentation Link."""
-    return {
-        s.studien_id: Link(url=s.dokumentation)
-        for s in synopse_studies
-        if s.dokumentation
-    }
-
-
-@pytest.fixture
 def keyword_text_by_study_id(
-    synopse_studies: list[SynopseStudy],
+    synopse_studies: list[MetadatenZuDatensaetzen],
 ) -> dict[str, list[Text]]:
     """Return a lookup from study ID to list of keyword Text."""
     return {
@@ -316,7 +275,9 @@ def keyword_text_by_study_id(
 
 
 @pytest.fixture
-def synopse_study(synopse_studies: list[SynopseStudy]) -> SynopseStudy:
+def synopse_study(
+    synopse_studies: list[MetadatenZuDatensaetzen],
+) -> MetadatenZuDatensaetzen:
     """Return a Synopse Study."""
     return synopse_studies[0]
 
@@ -331,17 +292,17 @@ def synopse_merged_organization_ids_by_query_string() -> dict[
 
 
 @pytest.fixture
-def synopse_projects() -> list[SynopseProject]:
+def synopse_projects() -> list[ProjektUndStudienverwaltung]:
     """Return a list of Synopse Projects."""
     return [
-        SynopseProject(
-            akronym_des_studientitels="BBCCDD_00",
+        ProjektUndStudienverwaltung(
+            studie="BBCCDD_00",
             anschlussprojekt="BBCCDD",
             beitragende="Roland Resolved",
-            beschreibung_der_studie="BBCCDD-Basiserhebung am RKI.",
-            externe_partner="Testpartner",
-            project_studientitel="Studie zu Lorem und Ipsum",
-            kontakt=["info@rki.de"],
+            BeschreibungStudie="BBCCDD-Basiserhebung am RKI.",
+            partner_extern="Testpartner",
+            projekt_studientitel="studie zu Lorem und Ipsum",
+            Kontakt=["info@rki.de"],
             projektbeginn="1999",
             projektdokumentation="""Z:\\Projekte\\Dokumentation
 
@@ -349,16 +310,16 @@ def synopse_projects() -> list[SynopseProject]:
 - Labor""",
             projektende="2000",
             studien_id="12345",
-            studienart_studientyp="Monitoring-Studie",
+            studien_art_typ="Monitoring-studie",
             verantwortliche_oe="C1",
-            interne_partner="fg99, C1",
+            partner_intern="fg99, C1",
         ),
-        SynopseProject(
-            akronym_des_studientitels="BBCCDD",
+        ProjektUndStudienverwaltung(
+            studie="BBCCDD",
             beitragende="Roland Resolved",
-            beschreibung_der_studie="BBCCDD-Basiserhebung am RKI.",
-            project_studientitel="Studie zu Lorem und Ipsum",
-            kontakt=["info@rki.de"],
+            BeschreibungStudie="BBCCDD-Basiserhebung am RKI.",
+            projekt_studientitel="studie zu Lorem und Ipsum",
+            Kontakt=["info@rki.de"],
             projektbeginn="1999",
             projektdokumentation="""Z:\\Projekte\\Dokumentation
 
@@ -366,14 +327,16 @@ def synopse_projects() -> list[SynopseProject]:
 - Labor""",
             projektende="2000",
             studien_id="12346",
-            studienart_studientyp="Monitoring-Studie",
+            studien_art_typ="Monitoring-studie",
             verantwortliche_oe="C1",
         ),
     ]
 
 
 @pytest.fixture
-def synopse_project(synopse_projects: list[SynopseProject]) -> SynopseProject:
+def synopse_project(
+    synopse_projects: list[ProjektUndStudienverwaltung],
+) -> ProjektUndStudienverwaltung:
     """Return a Synopse Project."""
     return synopse_projects[0]
 
@@ -382,7 +345,7 @@ def synopse_project(synopse_projects: list[SynopseProject]) -> SynopseProject:
 def extracted_activity() -> ExtractedActivity:
     """Return an extracted activity."""
     return ExtractedActivity(
-        abstract=[{"value": "Die Studie untersucht die Laune."}],
+        abstract=[{"value": "Die studie untersucht die Laune."}],
         activityType=["https://mex.rki.de/item/activity-type-6"],
         contact=[Identifier.generate(seed=123)],
         documentation=[
@@ -399,7 +362,7 @@ def extracted_activity() -> ExtractedActivity:
         shortName=[{"value": "BBCCDD_00"}],
         start=[TemporalEntity("2000")],
         theme=["https://mex.rki.de/item/theme-36"],
-        title=[Text(language=TextLanguage.DE, value="Studie zu Lorem und Ipsum")],
+        title=[Text(language=TextLanguage.DE, value="studie zu Lorem und Ipsum")],
     )
 
 
@@ -417,9 +380,9 @@ def resources_by_synopse_id(
 
 @pytest.fixture
 def extracted_variable_groups(
-    synopse_variables_by_thema: dict[str, list[SynopseVariable]],
+    synopse_variables_by_thema: dict[str, list[Variablenuebersicht]],
     resources_by_synopse_id: dict[str, ExtractedResource],
-    synopse_study_overviews: list[SynopseStudyOverview],
+    synopse_study_overviews: list[Datensatzuebersicht],
 ) -> list[ExtractedVariableGroup]:
     """Return a list of extracted variable groups."""
     return transform_synopse_variables_to_mex_variable_groups(